数据库重构&回复功能异常捕获

+++ src/main/java/com/example/demobs/controller/PublishController.java
@@ -19,7 +19,7 @@ public class PublishController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id,Model model){
+    public String edit(@PathVariable("id") Long id, Model model){
         QuestionDTO question = questionService.getQuestionById(id);
         model.addAttribute("title",question.getTitle());  //为了编辑
         model.addAttribute("description",question.getDescription());
@@ -37,7 +37,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id")Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title",title);  //为了回显
